Add only mapped recipes to mappedItems.recipes and ingredientIn
A recipe can be excluded because one of its items was
blacklisted/unknown/...
In this case, it will not be added to mappedRecipes.
However, items.ingredientIn and items.recipes was populated with a
recipe when it was part of the recipe's items regardless of the
validity of the rest of the recipe's items.

Example:
1. Recipe 'Test', input: Stone, output: Foo
2. Recipe 'Test' added to item Stone's ingredientIn
3. Recipe 'Test' excluded because output Foo is unknown
4. Stone has unknown recipe => problem

After this commit, recipes are only added to items when they are
considered valid.

This change also improves performance by:
- no longer looping over all recipes and their inputs/outputs for every
  item.
- avoiding an extra loop to check whether all recipe items are mapped.

Rough estimates seem to more than halve the execution time,
~1.8s -> ~0.8s.