Default to Cache-Control: no-cache
This marks resources as immediately stale and makes clients and proxies
validate the resource using If-Modified-Since and If-None-Match
headers.
The origin responds with either the content or 304 if the resource has
not changed and can be reused.

no-cache is the same as max-age=0,must-revalidate in modern times.
See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control#up-to-date_contents_always