Execute icon generation command using parallel
sync-icons now writes all commands to a file that can be executed using
gnu parallel.
Previously, sync-icon executed the commands itself which was.
It now takes 1m 26s on a AMD 7950x on WSL2.
The previous implementation took over 10min on WSL1 (ubuntu 20) but
1m 37s on WSL2 (arch).
The WSL version seems to matter more.

Also changed to use the "magick" command instead of "convert" which is
deprecated.